Skip rendering of the tooltip instead of hiding it
Using the `display: none` property, we won't render the tooltip when
it's not being showed (hover, click and so on).

However, the downside is that we won't be able to get the height of an
element that's set to `display: none`, so the visibility shuffle is done
to just calculate the size.

:thought_balloon: